Change ball_radius type to float.

+++ src/ball/ball.h
@@ -5,7 +5,7 @@
 
 typedef struct Ball {
   Vector2 position;
-  int radius;
+  float radius;
 } Ball;
 
 extern Ball ball;
